Fix:경험 선물, 예약 변경 수정
