+++ app/src/main/java/com/shall_we/dto/ExperienceMainRes.kt
@@ -1,7 +1,7 @@
 package com.shall_we.dto
 
 data class ExperienceMainRes(
-    val expCategories: List<ExpCategoryRes>,
+    val expCategoryRes: List<ExpCategoryRes>,
     val sttCategoryRes: List<SttCategoryRes>
 )
