+++ app/src/main/res/layout/fragment_changed_reservation.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".changeReservation.ChangedReservationFragment">
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/changed_reservation_layout"
+        android:backgroundTint="@color/white"
+        android:background="@color/white"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+    <ImageView
+        android:id="@+id/gifted"
+        android:layout_width="314dp"
+        android:layout_height="201dp"
+        android:layout_marginTop="100dp"
+        android:src="@drawable/gift"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_marginBottom="15dp"
+        android:layout_height="28dp"
+        android:text="예약 변경 성공"
+        android:textColor="#E31B54"
+        android:textSize="19dp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="@+id/gifted"
+        app:layout_constraintEnd_toEndOf="@+id/gifted"
+        app:layout_constraintStart_toStartOf="@+id/gifted"></TextView>
+
+
+    <TextView
+        android:id="@+id/gift_text01"
+        android:textSize="14dp"
+        android:textColor="@color/black"
+        android:text="쉘위이"
+        android:layout_marginTop="5dp"
+        app:layout_constraintTop_toBottomOf="@+id/gifted"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginLeft="40dp"
+        app:layout_constraintLeft_toLeftOf="@+id/gifted"></TextView>
+    <TextView
+        android:id="@+id/gift_text02"
+        android:textSize="14dp"
+        android:textColor="@color/black"
+        android:text="님께 함께 하고픈 마음을 전했어요"
+        android:layout_marginTop="5dp"
+        app:layout_constraintTop_toBottomOf="@+id/gifted"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintLeft_toRightOf="@+id/gift_text01"></TextView>
+
+    <Button
+        android:id="@+id/gift_btn01"
+        android:layout_width="328dp"
+        android:layout_height="39dp"
+        android:layout_marginTop="220dp"
+        android:background="@drawable/btn_design"
+        android:backgroundTint="#F15F79"
+        android:text="홈으로"
+        android:textColor="#FFFFFF"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/gift_text02">
+
+    </Button>
+
+    <Button
+
+        android:id="@+id/gift_btn02"
+        android:layout_width="328dp"
+        android:layout_height="39dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginBottom="50dp"
+        android:background="@drawable/btn_design"
+        android:backgroundTint="#FEF3F2"
+        android:text="보낸 선물함 확인하기"
+        android:textColor="#F15F79"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="@+id/gift_btn01"
+        app:layout_constraintTop_toBottomOf="@+id/gift_btn01">
+
+    </Button>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
