+++ app/src/main/res/layout/fragment_my_gift_received.xml
@@ -4,6 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:id="@+id/my_gift_recieved_layout"
     tools:context=".mypage.MyGiftReceivedFragment">
 
     <androidx.recyclerview.widget.RecyclerView
