+++ app/src/main/res/layout/fragment_mypage.xml
@@ -4,6 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/white"
+    android:backgroundTint="@color/white"
     android:id="@+id/mypage_layout"
     android:theme="@style/Theme.MaterialComponents"
     tools:context=".mypage.MypageFragment">
